+++ user/user.c
@@ -3,11 +3,17 @@
 
 extern char __stack_top[];
 
-int syscall(int sysno, int arg0, int arg1, int arg2) {
-    register int a0 __asm__("a0") = arg0;
-    register int a1 __asm__("a1") = arg1;
-    register int a2 __asm__("a2") = arg2;
-    register int a3 __asm__("a3") = sysno;
+/* ecall ABI (matches kernel/exception.c handle_syscall):
+ *   a3 = syscall number
+ *   a0..a2 = args
+ *   return: a0
+ * All slots are `long` so that 64-bit pointers survive lp64 without being
+ * truncated into 32-bit int. */
+long syscall(long sysno, long arg0, long arg1, long arg2) {
+    register long a0 __asm__("a0") = arg0;
+    register long a1 __asm__("a1") = arg1;
+    register long a2 __asm__("a2") = arg2;
+    register long a3 __asm__("a3") = sysno;
 
     __asm__ __volatile__("ecall"
                          : "=r"(a0)
@@ -18,24 +24,24 @@ int syscall(int sysno, int arg0, int arg1, int arg2) {
 }
 
 void putchar(char ch) {
-    syscall(SYS_PUTCHAR, ch, 0, 0);
+    syscall(SYS_PUTCHAR, (long) ch, 0, 0);
 }
 
 int getchar(void) {
-    return syscall(SYS_GETCHAR, 0, 0, 0);
+    return (int) syscall(SYS_GETCHAR, 0, 0, 0);
 }
 
 NORETURN void exit(void) {
     syscall(SYS_EXIT, 0, 0, 0);
-    for (;;); // Just in case!
+    for (;;);
 }
 
-int read(const char *filename, char *buf, int len) {
-    return syscall(SYS_READ, (int) filename, (int) buf, len);
+long read(const char *filename, char *buf, long len) {
+    return syscall(SYS_READ, (long) filename, (long) buf, len);
 }
 
-int write(const char *filename, const char *buf, int len) {
-    return syscall(SYS_WRITE, (int) filename, (int) buf, len);
+long write(const char *filename, const char *buf, long len) {
+    return syscall(SYS_WRITE, (long) filename, (long) buf, len);
 }
 
 NAKED
